idk wtf i'm doing

+++ src/services/claim-timer.js
@@ -12,6 +12,7 @@ module.exports = {
       Object.keys(instance.shared["timer"]).forEach((c) => {
         const e = instance.shared["timer"][c];
         e.forEach(async (s) => {
+          if (new Date() - s.last < 4000) continue;
           const e = await getTimer(s.time);
           if (!e) {
             s.msg.delete();
@@ -20,10 +21,17 @@ module.exports = {
             return;
           }
 
-          s.msg.edit(e);
+          s.msg.edit(e).then(() => {
+            const i = instance.shared["timer"][c].indexOf(s);
+            if (i !== -1)
+              instance.shared["timer"][c][i] = {
+                ...instance.shared["timer"][c][i],
+                last: new Date(),
+              };
+          });
         });
       });
-    }, 4000);
+    }, 2000);
   },
   stop: async (instance) => {
     if (updateInterval) clearInterval(updateInterval);
